refactor(cli): extract bare t3 eval full-suite callback to keep app.py under the module-health LOC cap
The #3222 backend wiring pushed src/teatree/cli/eval/app.py from 499 to 508
LOC, over the 500 module-health ceiling. Move the invoke_without_command
full-suite callback into full_suite_command.py (a distinct concern: the
whole-suite entry vs the targeted run command), dropping app.py to 448 LOC.
CLI surface unchanged; cli-reference stays byte-identical.
